refactor(test): Reorganize IR test suite into functional domain subdirectories
Major reorganization of the IR test suite for better maintainability and clarity:

Structural Changes:
- Organize tests into 8 functional subdirectories:
  * core/ - Core IR primitives (Span, Op, Var, Types)
  * expressions/ - Expression nodes (constants, binary ops, comparisons)
  * statements/ - Statement nodes (assign, loops, conditionals)
  * operators/ - Operator registry and operations
  * high_level/ - High-level constructs (Function, Program, Builder)
  * memory/ - Memory-related constructs (MemRef)
  * transforms/ - IR transformations (hash, equality, serialization)
  * printing/ - IR printing and visualization

File Changes:
- Split test_ir_base.py into focused test files in core/
- Split test_hash_equal.py (1410 lines) into:
  * transforms/test_hash.py (hash functionality)
  * transforms/test_equality.py (equality checks)
- Moved all existing test files to appropriate subdirectories
- Created __init__.py for each subdirectory

Documentation:
- Added comprehensive README.md with:
  * Directory structure overview
  * Test writing guidelines and naming conventions
  * Code examples and best practices
  * Commands for running tests
- All code comments and docstrings converted to English
