+++ tests/ut/ir/core/test_tuple_type.py
