+++ tests/ut/ir/expressions/test_binary_ops.py
@@ -7,7 +7,7 @@
 # See LICENSE in the root of the software repository for the full text of the License.
 # -----------------------------------------------------------------------------------------------------------
 
-"""Unit tests for PyPTO IR (Intermediate Representation) module."""
+"""Tests for binary operation expressions: Add, Sub, Mul, Div, Mod, etc."""
 
 from typing import cast
 
@@ -15,179 +15,8 @@ import pytest
 from pypto import DataType, ir
 
 
-class TestSpan:
-    """Tests for Span class."""
-
-    def test_span_creation(self):
-        """Test creating a Span with valid coordinates."""
-        span = ir.Span("test.py", 1, 5, 1, 10)
-        assert span.filename == "test.py"
-        assert span.begin_line == 1
-        assert span.begin_column == 5
-        assert span.end_line == 1
-        assert span.end_column == 10
-
-    def test_span_to_string(self):
-        """Test Span to_string method."""
-        span = ir.Span("test.py", 10, 20, 10, 30)
-        assert span.to_string() == "test.py:10:20"
-
-    def test_span_str_repr(self):
-        """Test Span __str__ and __repr__."""
-        span = ir.Span("example.py", 5, 1, 5, 10)
-        assert str(span) == "example.py:5:1"
-        assert repr(span) == "example.py:5:1"
-
-    def test_span_is_valid(self):
-        """Test Span is_valid method."""
-        valid_span = ir.Span("test.py", 1, 1, 1, 10)
-        assert valid_span.is_valid() is True
-
-        invalid_span = ir.Span("test.py", -1, -1, -1, -1)
-        assert invalid_span.is_valid() is False
-
-        # Test multi-line span
-        multiline_span = ir.Span("test.py", 1, 5, 3, 10)
-        assert multiline_span.is_valid() is True
-
-    def test_span_unknown(self):
-        """Test Span.unknown() factory method."""
-        span = ir.Span.unknown()
-        assert span.filename == ""
-        assert span.is_valid() is False
-
-    def test_span_immutability(self):
-        """Test that Span attributes are immutable."""
-        span = ir.Span.unknown()
-
-        # Attempting to modify should raise AttributeError
-        with pytest.raises(AttributeError):
-            span.filename = "other.py"  # type: ignore
-
-        with pytest.raises(AttributeError):
-            span.begin_line = 5  # type: ignore
-
-
-class TestOp:
-    """Tests for Op class."""
-
-    def test_op_creation(self):
-        """Test creating an Op."""
-        op = ir.Op("add")
-        assert op.name == "add"
-
-    def test_op_name_immutability(self):
-        """Test that Op name is immutable."""
-        op = ir.Op("multiply")
-        with pytest.raises(AttributeError):
-            op.name = "divide"  # type: ignore
-
-
-class TestVar:
-    """Tests for Var class."""
-
-    def test_var_creation(self):
-        """Test creating a Var expression."""
-        span = ir.Span("test.py", 1, 1, 1, 5)
-        var = ir.Var("x", ir.ScalarType(DataType.INT64), span)
-
-        assert var.name == "x"
-        assert var.span.filename == "test.py"
-
-    def test_var_is_expr(self):
-        """Test that Var is an instance of Expr."""
-        span = ir.Span("test.py", 1, 1, 1, 5)
-        var = ir.Var("x", ir.ScalarType(DataType.INT64), span)
-
-        assert isinstance(var, ir.Expr)
-        assert isinstance(var, ir.IRNode)
-
-    def test_var_immutability(self):
-        """Test that Var attributes are immutable."""
-        span = ir.Span("test.py", 1, 1, 1, 5)
-        var = ir.Var("x", ir.ScalarType(DataType.INT64), span)
-
-        # Attempting to modify should raise AttributeError
-        with pytest.raises(AttributeError):
-            var.name = "y"  # type: ignore
-
-
-class TestConstInt:
-    """Tests for ConstInt class."""
-
-    def test_const_creation(self):
-        """Test creating a ConstInt expression."""
-        span = ir.Span("test.py", 1, 1, 1, 5)
-        const = ir.ConstInt(42, DataType.INT64, span)
-
-        assert const.value == 42
-        assert const.span.filename == "test.py"
-
-    def test_const_integer(self):
-        """Test ConstInt with integer values."""
-        span = ir.Span("test.py", 1, 1, 1, 5)
-        const = ir.ConstInt(5, DataType.INT64, span)
-
-        assert const.value == 5
-
-    def test_const_is_expr(self):
-        """Test that ConstInt is an instance of Expr."""
-        span = ir.Span("test.py", 1, 1, 1, 5)
-        const = ir.ConstInt(10, DataType.INT64, span)
-
-        assert isinstance(const, ir.Expr)
-        assert isinstance(const, ir.IRNode)
-
-    def test_const_immutability(self):
-        """Test that ConstInt attributes are immutable."""
-        span = ir.Span("test.py", 1, 1, 1, 5)
-        const = ir.ConstInt(42, DataType.INT64, span)
-
-        # Attempting to modify should raise AttributeError
-        with pytest.raises(AttributeError):
-            const.value = 100  # type: ignore
-
-
-class TestConstBool:
-    """Tests for ConstBool class."""
-
-    def test_const_bool_creation_true(self):
-        """Test creating a ConstBool expression with True value."""
-        span = ir.Span("test.py", 1, 1, 1, 5)
-        const = ir.ConstBool(True, span)
-
-        assert const.value is True
-        assert const.span.filename == "test.py"
-        assert const.dtype == DataType.BOOL
-
-    def test_const_bool_creation_false(self):
-        """Test creating a ConstBool expression with False value."""
-        span = ir.Span("test.py", 1, 1, 1, 5)
-        const = ir.ConstBool(False, span)
-
-        assert const.value is False
-        assert const.dtype == DataType.BOOL
-
-    def test_const_bool_is_expr(self):
-        """Test that ConstBool is an instance of Expr."""
-        span = ir.Span("test.py", 1, 1, 1, 5)
-        const = ir.ConstBool(True, span)
-
-        assert isinstance(const, ir.Expr)
-        assert isinstance(const, ir.IRNode)
-
-    def test_const_bool_immutability(self):
-        """Test that ConstBool attributes are immutable."""
-        span = ir.Span("test.py", 1, 1, 1, 5)
-        const = ir.ConstBool(True, span)
-
-        # Attempting to modify should raise AttributeError
-        with pytest.raises(AttributeError):
-            const.value = False  # type: ignore
-
-
-class TestBinaryExpressions:
-    """Tests for binary expression classes."""
+class TestArithmeticOps:
+    """Tests for arithmetic binary operations."""
 
     def test_add_creation(self):
         """Test creating an Add expression."""
@@ -267,30 +96,33 @@ class TestBinaryExpressions:
         assert cast(ir.Var, floordiv_expr.left).name == "x"
         assert cast(ir.Var, floordiv_expr.right).name == "y"
 
-    def test_comparison_ops(self):
-        """Test creating comparison expressions."""
+    def test_pow(self):
+        """Test creating Pow expression."""
         span = ir.Span.unknown()
         dtype = DataType.INT64
         x = ir.Var("x", ir.ScalarType(dtype), span)
         y = ir.Var("y", ir.ScalarType(dtype), span)
 
-        eq_expr = ir.Eq(x, y, dtype, span)
-        assert cast(ir.Var, eq_expr.left).name == "x"
+        pow_expr = ir.Pow(x, y, dtype, span)
+        assert cast(ir.Var, pow_expr.left).name == "x"
+        assert cast(ir.Var, pow_expr.right).name == "y"
 
-        ne_expr = ir.Ne(x, y, dtype, span)
-        assert cast(ir.Var, ne_expr.left).name == "x"
+    def test_min_max(self):
+        """Test creating Min and Max expressions."""
+        span = ir.Span.unknown()
+        dtype = DataType.INT64
+        x = ir.Var("x", ir.ScalarType(dtype), span)
+        y = ir.Var("y", ir.ScalarType(dtype), span)
 
-        lt_expr = ir.Lt(x, y, dtype, span)
-        assert cast(ir.Var, lt_expr.left).name == "x"
+        min_expr = ir.Min(x, y, dtype, span)
+        assert cast(ir.Var, min_expr.left).name == "x"
 
-        le_expr = ir.Le(x, y, dtype, span)
-        assert cast(ir.Var, le_expr.left).name == "x"
+        max_expr = ir.Max(x, y, dtype, span)
+        assert cast(ir.Var, max_expr.left).name == "x"
 
-        gt_expr = ir.Gt(x, y, dtype, span)
-        assert cast(ir.Var, gt_expr.left).name == "x"
 
-        ge_expr = ir.Ge(x, y, dtype, span)
-        assert cast(ir.Var, ge_expr.left).name == "x"
+class TestLogicalOps:
+    """Tests for logical binary operations."""
 
     def test_logical_ops(self):
         """Test creating logical expressions."""
@@ -307,6 +139,10 @@ class TestBinaryExpressions:
 
         xor_expr = ir.Xor(x, y, dtype, span)
         assert cast(ir.Var, xor_expr.left).name == "x"
+
+
+class TestBitwiseOps:
+    """Tests for bitwise binary operations."""
 
     def test_bitwise_ops(self):
         """Test creating bitwise expressions."""
@@ -330,33 +166,9 @@ class TestBinaryExpressions:
         shr_expr = ir.BitShiftRight(x, y, dtype, span)
         assert cast(ir.Var, shr_expr.left).name == "x"
 
-    def test_min_max(self):
-        """Test creating Min and Max expressions."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        x = ir.Var("x", ir.ScalarType(dtype), span)
-        y = ir.Var("y", ir.ScalarType(dtype), span)
 
-        min_expr = ir.Min(x, y, dtype, span)
-        assert cast(ir.Var, min_expr.left).name == "x"
-
-        max_expr = ir.Max(x, y, dtype, span)
-        assert cast(ir.Var, max_expr.left).name == "x"
-
-    def test_pow(self):
-        """Test creating Pow expression."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        x = ir.Var("x", ir.ScalarType(dtype), span)
-        y = ir.Var("y", ir.ScalarType(dtype), span)
-
-        pow_expr = ir.Pow(x, y, dtype, span)
-        assert cast(ir.Var, pow_expr.left).name == "x"
-        assert cast(ir.Var, pow_expr.right).name == "y"
-
-
-class TestUnaryExpressions:
-    """Tests for unary expression classes."""
+class TestUnaryOps:
+    """Tests for unary operations."""
 
     def test_neg_creation(self):
         """Test creating a Neg expression."""
@@ -497,7 +309,7 @@ class TestNestedExpressions:
 
 
 class TestImmutability:
-    """Tests for immutability of IR nodes."""
+    """Tests for immutability of expression nodes."""
 
     def test_expr_operands_immutable(self):
         """Test that binary expression operands cannot be modified."""
@@ -510,43 +322,6 @@ class TestImmutability:
         # Attempting to modify should raise AttributeError
         with pytest.raises(AttributeError):
             add_expr.left = ir.Var("z", ir.ScalarType(dtype), span)  # type: ignore
-
-    def test_span_immutable_in_node(self):
-        """Test that span attribute in IRNode is immutable."""
-        span = ir.Span.unknown()
-        dtype = DataType.INT64
-        x = ir.Var("x", ir.ScalarType(dtype), span)
-
-        # Attempting to modify should raise AttributeError
-        with pytest.raises(AttributeError):
-            x.span = ir.Span("other.py", 2, 2, 2, 5)  # type: ignore
-
-
-class TestSpanTracking:
-    """Tests for source location tracking via Span."""
-
-    def test_span_preserved_in_tree(self):
-        """Test that spans are preserved throughout the expression tree."""
-        span1 = ir.Span("file1.py", 1, 1, 1, 5)
-        span2 = ir.Span("file2.py", 2, 2, 2, 5)
-        span3 = ir.Span("file3.py", 3, 3, 3, 10)
-        dtype = DataType.INT64
-
-        x = ir.Var("x", ir.ScalarType(dtype), span1)
-        y = ir.Var("y", ir.ScalarType(dtype), span2)
-        add_expr = ir.Add(x, y, dtype, span3)
-
-        assert x.span.filename == "file1.py"
-        assert y.span.filename == "file2.py"
-        assert add_expr.span.filename == "file3.py"
-
-    def test_unknown_span(self):
-        """Test creating IR nodes with unknown spans."""
-        unknown_span = ir.Span.unknown()
-        x = ir.Var("x", ir.ScalarType(DataType.INT64), unknown_span)
-
-        assert x.span.is_valid() is False
-        assert x.span.filename == ""
 
 
 if __name__ == "__main__":
