+++ tests/ut/ir/operators/test_op_registry.py
