+++ tests/ut/ir/operators/test_tensor_ops.py
